% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_clips.R
\name{concat_clips}
\alias{concat_clips}
\title{Concatenate GoPro video clips}
\usage{
concat_clips(folder, no_audio = TRUE)
}
\arguments{
\item{folder}{A Character providing the path to the folder containing the video clips to stich together}

\item{no_audio}{A logical dropping (TRUE) or keeping audio (FALSE)}
}
\description{
Concatenate GoPro video clips
}
\examples{
\dontrun{
concat_clips("09_16_23 1108 baseline", no_audio=TRUE)
}
}
