% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_clips.R
\name{concat_clips}
\alias{concat_clips}
\title{Concatenate GoPro video clips}
\usage{
concat_clips(
  folder_in,
  folder_out = folder_in,
  full_video_name,
  no_audio = TRUE
)
}
\arguments{
\item{folder_in}{A Character providing the path to the folder containing the video clips to stitch together}

\item{folder_out}{A Character providing the path to the folder where to save the concatenated video}

\item{full_video_name}{A Character filename for the full length video}

\item{no_audio}{A logical dropping (TRUE) or keeping audio (FALSE)}
}
\value{
A Character path to stitched video
}
\description{
Concatenate GoPro video clips
}
\examples{
\dontrun{
concat_clips("folder_in=baseline", full_video_name="baseline_full.mp4", no_audio=TRUE)
}
}
